Add a ufw rule to allow routed/forwarded packets
CIS hardening demands that the default policy for routed packets be
deny, but the rule being added allows all routed packets as if the
default policy were allow.  I am following the letter of the law, but
not the intent.

Soon I will have billboards up and down the local interstates
advertising my ambulance-chasing services.  Ugh.