Add AWS_MAX_ATTEMPTS env variable to build workflow
Since we are using the default value of 15 seconds for AWS_POLL_DELAY_SECONDS, this corresponds to a timeout of an hour and a half.  We need such a long timeout because it can take almost an hour to create this AMI.

This is already present in the prerelease and release workflows.  It should have been done in #119, but it was accidentally overlooked.